moves the useHighlight hook
